feat: include path for duplicate key exception
