+++ src/getJsonPathForPosition.ts
@@ -1,5 +1,7 @@
-import { GetJsonPathForPosition, JsonPath } from '@stoplight/types';
-import { Kind, YamlMap, YAMLMapping, YAMLNode, YAMLScalar, YAMLSequence } from 'yaml-ast-parser';
+import { GetJsonPathForPosition } from '@stoplight/types';
+import { Kind, YamlMap, YAMLMapping, YAMLNode, YAMLSequence } from 'yaml-ast-parser';
+import { buildJsonPath } from './buildJsonPath';
+import { isValidNode } from './utils';
 
 export const getJsonPathForPosition: GetJsonPathForPosition<YAMLNode, number[]> = (
   { ast, lineMap },
@@ -18,8 +20,6 @@ export const getJsonPathForPosition: GetJsonPathForPosition<YAMLNode, number[]> 
   if (path.length === 0) return;
   return path;
 };
-
-const isValidNode = (node: YAMLNode) => node !== null && node !== undefined;
 
 function* walk(node: YAMLNode): IterableIterator<YAMLNode> {
   switch (node.kind) {
@@ -114,47 +114,4 @@ function findClosestScalar(container: YAMLNode, offset: number, line: number, li
   }
 
   return container;
-}
-
-function buildJsonPath(node: YAMLNode) {
-  const path: JsonPath = [];
-
-  let prevNode: YAMLNode = node;
-
-  while (node) {
-    switch (node.kind) {
-      case Kind.SCALAR:
-        path.unshift((node as YAMLScalar).value);
-        break;
-      case Kind.MAPPING:
-        if (prevNode !== (node as YAMLMapping).key) {
-          if (
-            path.length > 0 &&
-            isValidNode((node as YAMLMapping).value) &&
-            (node as YAMLMapping).value.value === path[0]
-          ) {
-            path[0] = (node as YAMLMapping).key.value;
-          } else {
-            path.unshift((node as YAMLMapping).key.value);
-          }
-        }
-        break;
-      case Kind.SEQ:
-        if (prevNode) {
-          const index = (node as YAMLSequence).items.indexOf(prevNode);
-          if (prevNode.kind === Kind.SCALAR) {
-            path[0] = index;
-            // always better to point to parent node rather than nothing
-          } else if (index !== -1) {
-            path.unshift(index);
-          }
-        }
-        break;
-    }
-
-    prevNode = node;
-    node = node.parent;
-  }
-
-  return path;
 }
