+++ src/utils.ts
@@ -1,0 +1,3 @@
+import { YAMLNode } from 'yaml-ast-parser';
+
+export const isValidNode = (node: YAMLNode) => node !== null && node !== undefined;
